Move payment submission files to new app/submit-payment dir
Create new directory to be more consistent with existing dir naming.
