+++ app/setup-payment/index.js
@@ -1,5 +1,3 @@
 const { paymentAuthoriseConsent } = require('./payment-authorise-consent');
-const { paymentSubmission } = require('./payment-submission');
 
 exports.paymentAuthoriseConsent = paymentAuthoriseConsent;
-exports.paymentSubmission = paymentSubmission;
