+++ app/submit-payment/index.js
@@ -1,0 +1,3 @@
+const { paymentSubmission } = require('./payment-submission');
+
+exports.paymentSubmission = paymentSubmission;
